Introduce utils file
