+++ .gitignore
@@ -1,1 +1,1 @@
-*.out
+main
