+++ main.c
@@ -2,6 +2,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include "gadgets.h"
+#include "utils.h"
 
 typedef struct Var {
     char* name;
@@ -21,26 +22,16 @@ int main(){
     for (int i = 0 ; i < progLines ; i++){
         char* line = strdup(prog[i]);
         char* opcode = strtok(line, " ");
-        // Consider refactoring to function
         char* operands = strtok(NULL, "");  // Save all operands
-        char* operand = strtok(operands, ",");
         char** operandList = malloc(3*20*sizeof(char));  // Max 3 operands at 20 chars each
-        int numOperands = 0;
-        // Iterate though operands 
-        while(operand != NULL){
-            removeChars(operand,' ');
-            operandList[numOperands] = operand;
-            numOperands++;
-            operand = strtok(NULL, ",");
-        }
+        int numOperands = getOperands(operandList, operands);
         
         if(strcmp(opcode,"Var") == 0){
             Var newVar = {
                 operandList[0],
                 atoi(operandList[1])
             };
-            vars[i] = newVar;
-              // Use hastable? Also probs shouldn't save value yet??
+            vars[i] = newVar;  // Use hastable? Also probs shouldn't save value yet??
         }
     }
     printf("%d %d\n",vars[0].value,vars[1].value);
