+++ utils.h
@@ -1,0 +1,3 @@
+void removeChars(char* str, char c);
+
+int getOperands(char** operandList, char* operandString);
